+++ predicators/ground_truth_models/doors/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for doors environment and variants."""
+
+from .nsrts import DoorsGroundTruthNSRTFactory
+
+__all__ = ["DoorsGroundTruthNSRTFactory"]
